refactor: changes on modules
