+++ services/video_streaming/app/config/config.go
@@ -1,4 +1,4 @@
-package initializers
+package config
 
 import (
 	"net/http"
