+++ services/video_streaming/app/library/aws_helpers/get_resolver.go
@@ -1,7 +1,8 @@
-package library
+package aws_helpers
 import(
 	"github.com/aws/aws-sdk-go/aws/endpoints"
 )
+
 func GetResolver(url string) endpoints.ResolverFunc {
 	defaultResolver := endpoints.DefaultResolver()
