+++ services/video_streaming/app/library/aws_helpers/get_session.go
@@ -1,4 +1,4 @@
-package library
+package aws_helpers
 
 import (
 	"github.com/aws/aws-sdk-go/aws"
