+++ services/video_streaming/app/library/aws_helpers/go.mod
@@ -1,0 +1,7 @@
+module aws_helpers
+
+go 1.20
+
+require github.com/aws/aws-sdk-go v1.44.297
+
+require github.com/jmespath/go-jmespath v0.4.0 // indirect
