+++ services/video_streaming/app/models/go.mod
@@ -1,0 +1,7 @@
+module models
+
+go 1.20
+
+require github.com/aws/aws-sdk-go v1.44.297
+
+require github.com/jmespath/go-jmespath v0.4.0 // indirect
