+++ services/video_streaming/app/resolvers/base_resolver.go
@@ -1,0 +1,5 @@
+package resolvers
+
+type Resolver struct {
+	Payload interface{}
+}
