+++ services/video_streaming/server.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"net/http"
 
+	. "config"
 	. "router"
 	. "initializers"
 )
